Moved the sources router into its own package.
