+++ pkg/config/config.go
@@ -1,12 +1,10 @@
 package config
 
 import (
-	"context"
 	_ "embed"
 	"fmt"
 	"os"
 	"path/filepath"
-	"reflect"
 	"regexp"
 	"strings"
 	"time"
@@ -16,16 +14,9 @@ import (
 	"github.com/knadh/koanf/providers/env"
 	"github.com/knadh/koanf/providers/file"
 	"github.com/knadh/koanf/providers/rawbytes"
-	"github.com/kubeshop/botkube/pkg/utils"
-	"github.com/sirupsen/logrus"
 	"github.com/spf13/pflag"
 	"golang.org/x/text/cases"
 	"golang.org/x/text/language"
-	coreV1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/meta"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/client-go/dynamic"
-	"k8s.io/client-go/tools/cache"
 )
 
 //go:embed default.yaml
@@ -135,337 +126,6 @@ type Config struct {
 
 	Analytics Analytics `yaml:"analytics"`
 	Settings  Settings  `yaml:"settings"`
-}
-
-type mergedEvents map[string]map[EventType]struct{}
-
-type SourcesRouter struct {
-	log        logrus.FieldLogger
-	mapper     meta.RESTMapper
-	dynamicCli dynamic.Interface
-	table      map[string][]RoutedEvent
-	bindings   map[string]struct{}
-	informers  map[string]Informer
-}
-
-type SourceRoutes struct {
-	source     string
-	namespaces []string
-}
-
-type RoutedEvent struct {
-	event  EventType
-	routes []SourceRoutes
-}
-
-type Informer struct {
-	informer        cache.SharedIndexInformer
-	log             logrus.FieldLogger
-	mapper          meta.RESTMapper
-	dynamicCli      dynamic.Interface
-	events          []EventType
-	mappedResources []string
-	mappedEvent     EventType
-}
-
-func (i Informer) handleRouted(ctx context.Context, resource string, target EventType, sourceRoutes []SourceRoutes, fn eventHandler) {
-	switch target {
-	case CreateEvent:
-		i.informer.AddEventHandler(cache.ResourceEventHandlerFuncs{
-			AddFunc: func(obj interface{}) {
-				if sources := sourcesForObjNamespace(ctx, sourceRoutes, obj, i.log, i.mapper, i.dynamicCli); len(sources) > 0 {
-					fn(ctx, resource, sources)(obj, nil)
-				}
-			},
-		})
-	case DeleteEvent:
-		i.informer.AddEventHandler(cache.ResourceEventHandlerFuncs{
-			DeleteFunc: func(obj interface{}) {
-				if sources := sourcesForObjNamespace(ctx, sourceRoutes, obj, i.log, i.mapper, i.dynamicCli); len(sources) > 0 {
-					fn(ctx, resource, sources)(obj, nil)
-				}
-			},
-		})
-	case UpdateEvent:
-		i.informer.AddEventHandler(cache.ResourceEventHandlerFuncs{
-			UpdateFunc: func(oldObj, newObj interface{}) {
-				if sources := sourcesForObjNamespace(ctx, sourceRoutes, newObj, i.log, i.mapper, i.dynamicCli); len(sources) > 0 {
-					fn(ctx, resource, sources)(newObj, oldObj)
-				}
-			},
-		})
-	}
-}
-
-func (i Informer) handleMapped(ctx context.Context, targetEvent EventType, routeTable map[string][]RoutedEvent, fn eventHandler) {
-	i.informer.AddEventHandler(cache.ResourceEventHandlerFuncs{
-		AddFunc: func(obj interface{}) {
-			var eventObj coreV1.Event
-			err := utils.TransformIntoTypedObject(obj.(*unstructured.Unstructured), &eventObj)
-			if err != nil {
-				i.log.Errorf("Unable to transform object type: %v, into type: %v", reflect.TypeOf(obj), reflect.TypeOf(eventObj))
-			}
-			_, err = cache.MetaNamespaceKeyFunc(obj)
-			if err != nil {
-				i.log.Errorf("Failed to get MetaNamespaceKey from event resource")
-				return
-			}
-
-			// Find involved object type
-			gvr, err := utils.GetResourceFromKind(i.mapper, eventObj.InvolvedObject.GroupVersionKind())
-			if err != nil {
-				i.log.Errorf("Failed to get involved object: %v", err)
-				return
-			}
-
-			if !i.canHandleEvent(eventObj.Type) {
-				return
-			}
-
-			gvrToString := utils.GVRToString(gvr)
-			if !i.includesSrcResource(gvrToString) {
-				return
-			}
-
-			sourceRoutes := sourceRoutes(routeTable, gvrToString, targetEvent)
-			sources := sourcesForObjNamespace(ctx, sourceRoutes, obj, i.log, i.mapper, i.dynamicCli)
-			if len(sources) == 0 {
-				return
-			}
-			fn(ctx, gvrToString, sources)(obj, nil)
-		},
-	})
-}
-
-func sourcesForObjNamespace(ctx context.Context, routes []SourceRoutes, obj interface{}, log logrus.FieldLogger, mapper meta.RESTMapper, cli dynamic.Interface) []string {
-	var out []string
-
-	objectMeta, err := utils.GetObjectMetaData(ctx, cli, mapper, obj)
-	if err != nil {
-		log.Errorf("while getting object metadata: %s", err.Error())
-		return []string{}
-	}
-
-	targetNs := objectMeta.Namespace
-	for _, route := range routes {
-		for _, ns := range route.namespaces {
-			if ns == targetNs || ns == "all" {
-				out = append(out, route.source)
-				break
-			}
-		}
-	}
-
-	return out
-}
-
-func (i Informer) canHandleEvent(target string) bool {
-	for _, e := range i.events {
-		if strings.ToLower(target) == e.String() {
-			return true
-		}
-	}
-	return false
-}
-
-func (i Informer) includesSrcResource(resource string) bool {
-	for _, src := range i.mappedResources {
-		if src == resource {
-			return true
-		}
-	}
-	return false
-}
-
-func (r *SourcesRouter) AddAnySlackBindings(c IdentifiableMap[ChannelBindingsByName]) {
-	for _, name := range c {
-		for _, source := range name.Bindings.Sources {
-			r.bindings[source] = struct{}{}
-		}
-	}
-}
-
-func (r *SourcesRouter) BuildTable(cfg *Config) *SourcesRouter {
-	sources := r.GetBoundSources(cfg.Sources)
-	mergedEvents := mergeResourceEvents(sources)
-
-	for resource, resourceEvents := range mergedEvents {
-		eventRoutes := mergeEventRoutes(resource, sources)
-		fmt.Printf("\n\nCOLLECTED_ROUTES: %+v\n\n", eventRoutes)
-		r.buildTable(resource, resourceEvents, eventRoutes)
-	}
-
-	return r
-}
-
-func mergeResourceEvents(sources IndexableMap[Sources]) mergedEvents {
-	out := map[string]map[EventType]struct{}{}
-	for _, srcGroupCfg := range sources {
-		for _, resource := range srcGroupCfg.Kubernetes.Resources {
-			if _, ok := out[resource.Name]; !ok {
-				out[resource.Name] = make(map[EventType]struct{})
-			}
-			for _, e := range flattenEvents(resource.Events) {
-				out[resource.Name][e] = struct{}{}
-			}
-		}
-	}
-	return out
-}
-
-func mergeEventRoutes(resource string, sources IndexableMap[Sources]) map[EventType][]SourceRoutes {
-	out := make(map[EventType][]SourceRoutes)
-	for srcGroupName, srcGroupCfg := range sources {
-		for _, r := range srcGroupCfg.Kubernetes.Resources {
-			for _, e := range flattenEvents(r.Events) {
-				if resource == r.Name {
-					out[e] = append(out[e], SourceRoutes{source: srcGroupName, namespaces: r.Namespaces.Include})
-				}
-			}
-		}
-	}
-	return out
-}
-
-func (r *SourcesRouter) buildTable(resource string, events map[EventType]struct{}, pairings map[EventType][]SourceRoutes) {
-	for evt := range events {
-		if _, ok := r.table[resource]; !ok {
-
-			r.table[resource] = []RoutedEvent{{
-				event:  evt,
-				routes: pairings[evt],
-			}}
-
-		} else {
-			r.table[resource] = append(r.table[resource], RoutedEvent{event: evt, routes: pairings[evt]})
-		}
-	}
-}
-
-func (r *SourcesRouter) GetBoundSources(sources IndexableMap[Sources]) IndexableMap[Sources] {
-	out := make(IndexableMap[Sources])
-	for name, t := range sources {
-		if _, ok := r.bindings[name]; ok {
-			out[name] = t
-		}
-	}
-	return out
-}
-
-type registrationHandler func(resource string) cache.SharedIndexInformer
-type eventHandler func(ctx context.Context, resource string, sources []string) func(obj, oldObj interface{})
-
-func (r *SourcesRouter) RegisterRoutedInformers(targetEvents []EventType, handler registrationHandler) {
-	resources := r.resourcesForEvents(targetEvents)
-	for _, resource := range resources {
-		r.informers[resource] = Informer{
-			informer:   handler(resource),
-			events:     r.resourceEvents(resource),
-			log:        r.log,
-			mapper:     r.mapper,
-			dynamicCli: r.dynamicCli,
-		}
-	}
-}
-
-func (r *SourcesRouter) RegisterMappedInformer(srcEvent EventType, dstEvent EventType, handler registrationHandler) {
-	srcResources := r.resourcesForEvents([]EventType{srcEvent})
-	if len(srcResources) == 0 {
-		return
-	}
-
-	dstResource := "v1/events"
-	r.informers[dstResource] = Informer{
-		informer:        handler(dstResource),
-		events:          []EventType{dstEvent},
-		mappedResources: srcResources,
-		mappedEvent:     srcEvent,
-		log:             r.log,
-		mapper:          r.mapper,
-		dynamicCli:      r.dynamicCli}
-}
-
-func (r *SourcesRouter) HandleRoutedEvent(ctx context.Context, target EventType, handlerFn eventHandler) {
-	for resource, informer := range r.informers {
-		if informer.canHandleEvent(target.String()) {
-			sourceRoutes := r.sourceRoutes(resource, target)
-			informer.handleRouted(ctx, resource, target, sourceRoutes, handlerFn)
-		}
-	}
-}
-
-func (r *SourcesRouter) HandleMappedEvent(ctx context.Context, targetEvent EventType, handlerFn eventHandler) {
-	if informer, ok := r.mappedInformer(targetEvent); ok {
-		informer.handleMapped(ctx, targetEvent, r.table, handlerFn)
-	}
-}
-
-func (r *SourcesRouter) sourceRoutes(resource string, targetEvent EventType) []SourceRoutes {
-	return sourceRoutes(r.table, resource, targetEvent)
-}
-
-func sourceRoutes(routeTable map[string][]RoutedEvent, targetResource string, targetEvent EventType) []SourceRoutes {
-	var out []SourceRoutes
-	for _, routedEvent := range routeTable[targetResource] {
-		if routedEvent.event == targetEvent {
-			out = append(out, routedEvent.routes...)
-		}
-	}
-	return out
-}
-
-func (r *SourcesRouter) resourceEvents(resource string) []EventType {
-	var out []EventType
-	for _, routedEvent := range r.table[resource] {
-		out = append(out, routedEvent.event)
-	}
-	return out
-}
-
-func (r *SourcesRouter) resourcesForEvents(targets []EventType) []string {
-	var out []string
-	for _, target := range targets {
-		for resource, routedEvents := range r.table {
-			for _, routedEvent := range routedEvents {
-				if routedEvent.event == target {
-					out = append(out, resource)
-					break
-				}
-			}
-		}
-	}
-	return out
-}
-
-func (r *SourcesRouter) mappedInformer(event EventType) (Informer, bool) {
-	for _, informer := range r.informers {
-		if informer.mappedEvent == event {
-			return informer, true
-		}
-	}
-	return Informer{}, false
-}
-
-func flattenEvents(events []EventType) []EventType {
-	var out []EventType
-	for _, event := range events {
-		if event == AllEvent {
-			out = append(out, []EventType{CreateEvent, UpdateEvent, DeleteEvent, ErrorEvent}...)
-		} else {
-			out = append(out, event)
-		}
-	}
-	return out
-}
-
-func NewSourcesRouter(mapper meta.RESTMapper, dynamicCli dynamic.Interface, log logrus.FieldLogger) *SourcesRouter {
-	return &SourcesRouter{
-		log:        log,
-		mapper:     mapper,
-		dynamicCli: dynamicCli,
-		table:      make(map[string][]RoutedEvent),
-		informers:  make(map[string]Informer),
-	}
 }
 
 // ChannelBindingsByName contains configuration bindings per channel.
